Change the plugins needed by bootstrax.
Boostrax currently crashes as it does not have the daqreader plugin. Hence, I propose to add all that is needed to go up to the default of ''--target'' (currently ''event_info''). 

We might also opt to change this default since 'high level' changes in e.g. event building cause bootstrax to fail (which we don’t want) whereas they are not necessarily the type of data what we intent to put into Rocio.