Update peak_processing.py
Implementation of a new Peak classification algorithm similar to the one used in pax and based on the rise time/ tight coincidence variables of the peak. I added two new variables in the "PeakBasics" plugin (rise_time and range_90p_area) and a new plugin "TightCoincidence" to compute the tight coincidence variable. Both are used in the new version of the "PeakClassification" plugin.