Update led_calibration.py
There were still the XENON1T default option. Update to current one.
I have removed 'dt' and 'length' since are not usefull.
For the moment I keapt the amplitude in the noise window even if we are not useing anymore. We can still use to make a fast comparison of baseline between the noise run and the led run.