Small tweak mongo config storage
So this was slightly confusing. Imagine someone wants (for some reason) upload the same file twice under a different name (e.g. because it is a placeholder). Then we also need to check that the file exists under that name. Otherwise it will appear like there is file missing from the database.