Fix strax version
We should have already done this directly after merging https://github.com/AxFoundation/strax/pull/367 to reflect the change in the low-level processing in straxen. This change is reflected in a version bump in the PulseProcessing but one should not mix an older strax with this newer straxen. The change in the base lining is than not reflected but nothing is stopping one from trying to process data with it.